backend/pyproject.toml
[tool.poetry]
name = "mayim"
version = "0.1.0"
description = "A project for managing Bible translations and references"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "backend" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.23.2"
sqlalchemy = "^1.4.41"
sqlmodel = "^0.0.6"
pydantic = "<2"
python-dotenv = "^0.21.0"
justpy = "^0.4.0"
langchain = "^0.0.30"
langchain-cli = ">=0.0.15"
langchain-ollama = "^0.1.0"
typing-extensions = "^4.3.0"
websockets = "^10.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"